spring-data-jdbc flyway

+++ liquibase/src/main/java/dbmigrations/liquibase/domain/repositories/DatabaseChangelogLockRepository.java
@@ -2,8 +2,8 @@ package dbmigrations.liquibase.domain.repositories;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import dbmigrations.liquibase.domain.model.DatabaseChangelogLock;
+import dbmigrations.liquibase.domain.models.DatabaseChangelogLock;
 
-public interface DatabaseChangelogLockRepository extends PagingAndSortingRepository<DatabaseChangelogLock,Integer> {
+public interface DatabaseChangelogLockRepository extends PagingAndSortingRepository<DatabaseChangelogLock,Long> {
     
 }
